third_party/blink/web_tests/external/wpt/pointerevents/compat/pointerevent_mouseevent_key_pressed-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer events correctly show Alt pressed
  promise_test: Unhandled rejection with value: object "Error: We do not support keydown and keyup mixed with other actions, please use test_driver.send_keys. See crbug.com/893480."
[FAIL] Pointer events correctly show Control pressed
  promise_test: Unhandled rejection with value: object "Error: We do not support keydown and keyup mixed with other actions, please use test_driver.send_keys. See crbug.com/893480."
[FAIL] Pointer events correctly show Meta pressed
  promise_test: Unhandled rejection with value: object "Error: We do not support keydown and keyup mixed with other actions, please use test_driver.send_keys. See crbug.com/893480."
[FAIL] Pointer events correctly show Shift pressed
  promise_test: Unhandled rejection with value: object "Error: We do not support keydown and keyup mixed with other actions, please use test_driver.send_keys. See crbug.com/893480."
[FAIL] Pointer events correctly show Alt,Control,Meta,Shift pressed
  promise_test: Unhandled rejection with value: object "Error: We do not support keydown and keyup mixed with other actions, please use test_driver.send_keys. See crbug.com/893480."
Harness: the test ran to completion.
